testing files for backend and frontend with jest and React Testing Library
