+++ backend/react-devtools.d.ts
@@ -1,0 +1,3 @@
+interface Window {
+  __REACT_DEVTOOLS_GLOBAL_HOOK__?: any;
+}
