+++ package.json
@@ -5,7 +5,8 @@
   "main": "index.js",
   "scripts": {
     "dev": "webpack serve --open --hot --env NODE_ENV=development",
-    "build": "webpack --env NODE_ENV=production"
+    "build": "webpack --env NODE_ENV=production",
+    "test": "jest"
   },
   "keywords": [],
   "author": "",
@@ -24,12 +25,22 @@
   },
   "devDependencies": {
     "@babel/core": "^7.21.8",
-    "@babel/preset-env": "^7.21.5",
+    "@babel/preset-env": "^7.22.2",
     "@babel/preset-react": "^7.18.6",
+    "@babel/preset-typescript": "^7.21.5",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
     "@types/chrome": "^0.0.236",
+    "@types/jest": "^29.5.1",
+    "@types/react": "^18.2.7",
+    "babel-jest": "^29.5.0",
     "babel-loader": "^9.1.2",
     "copy-webpack-plugin": "^11.0.0",
     "html-webpack-plugin": "^5.5.1",
+    "jest": "^29.5.0",
+    "jest-environment-jsdom": "^29.5.0",
+    "jsdom": "^22.0.0",
+    "ts-jest": "^29.1.0",
     "webpack": "^5.82.0",
     "webpack-cli": "^5.1.0",
     "webpack-dev-server": "^4.15.0"
