+++ testing-files/FiberNode.js
@@ -1,0 +1,225 @@
+// a server-side component
+const subNode5 = {
+    actualDuration: 0.39999985694885254,
+    actualStartTime: 2888.99000999046326,
+    alternate: null,
+    child: null,
+    childLanes: 0,
+    deletions: null,
+    dependencies: null,
+    elementType: {$$typeof: 'Symbol(react.lazy)',
+                  _payload: {
+                    reason: null,
+                    status: "fullfilled",
+                    value: {
+                        length: 0,
+                        name: "A SSC"
+                    }
+                  }},
+    flags: 0,
+    index: 0,
+    key: null,
+    lanes: 0,
+    memoizedProps: null,
+    memoizedState: null,
+    mode: 3,
+    pendingProps: null,
+    ref: null,
+    refCleanup: null,
+    return: null,
+    selfBaseDuration: 0.09999990463256836,
+    sibling: null,
+    stateNode: null,
+    subtreeFlags: 11536897,
+    tag: 0,
+    treeBaseDuration: 5.699999809265137,
+    type: null,
+    updateQueue: null,
+    _debugHookTypes: null,
+    _debugNeedsRemount: false,
+    _debugOwner: null,
+    _debugSource: null,
+}
+
+// a client-side component
+const subNode4 = {
+    actualDuration: 0.6099999046325684,
+    actualStartTime: 2888.49000999046326,
+    alternate: null,
+    child: null,
+    childLanes: 0,
+    deletions: null,
+    dependencies: null,
+    elementType: {length: 1,
+                  name: "A CSC",
+                },
+    flags: 8388608,
+    index: 0,
+    key: null,
+    lanes: 0,
+    memoizedProps: {},
+    memoizedState: {baseQueue: null,
+                    baseState: "",
+                    memoizedState: ""},
+    mode: 3,
+    pendingProps: null,
+    ref: null,
+    refCleanup: null,
+    return: null,
+    selfBaseDuration: 0.09999990463256836,
+    sibling: null,
+    stateNode: null,
+    subtreeFlags: 14682117,
+    tag: 0,
+    treeBaseDuration: 5.699999809265137,
+    type: null,
+    updateQueue: null,
+    _debugHookTypes: ['useState', 'useState', 'useEffect'],
+    _debugNeedsRemount: false,
+    _debugOwner: null,
+    _debugSource: null,
+}
+
+// an h1 element
+const subNode3 = {
+    actualDuration: 0.5999999046325684,
+    actualStartTime: 2887.99000999046326,
+    alternate: null,
+    child: subNode5,
+    childLanes: 0,
+    deletions: null,
+    dependencies: null,
+    elementType: "h1",
+    flags: 0,
+    index: 0,
+    key: null,
+    lanes: 0,
+    memoizedProps: {children: 'Welcome To My Application'},
+    memoizedState: null,
+    mode: 3,
+    pendingProps: {children: 'Welcome To My Application'},
+    ref: null,
+    refCleanup: null,
+    return: null,
+    selfBaseDuration: 0.09999990463256836,
+    sibling: subNode4,
+    stateNode: null,
+    subtreeFlags: 0,
+    tag: 5,
+    treeBaseDuration: 5.699999809265137,
+    type: null,
+    updateQueue: null,
+    _debugHookTypes: null,
+    _debugNeedsRemount: false,
+    _debugOwner: null,
+    _debugSource: null,
+}
+
+const subNode2 = {
+    actualDuration: 0.5999999046325684,
+    actualStartTime: 2887.8999999046326,
+    alternate: null,
+    child: subNode3,
+    childLanes: 0,
+    deletions: null,
+    dependencies: null,
+    elementType: {$$typeof: 'Symbol(react.provider)',
+                  Consumer: "",
+                  Provider: "",
+                  disPlayName: "HeadManagerContext",
+                  _currentRenderer: null},
+    flags: 0,
+    index: 0,
+    key: null,
+    lanes: 0,
+    memoizedProps: null,
+    memoizedState: null,
+    mode: 3,
+    pendingProps: null,
+    ref: null,
+    refCleanup: null,
+    return: null,
+    selfBaseDuration: 0.09999990463256836,
+    sibling: null,
+    stateNode: null,
+    subtreeFlags: 14682117,
+    tag: 10,
+    treeBaseDuration: 5.699999809265137,
+    type: null,
+    updateQueue: null,
+    _debugHookTypes: null,
+    _debugNeedsRemount: false,
+    _debugOwner: null,
+    _debugSource: null,
+}
+
+const subNode = {
+    actualDuration: 0.5999999046325684,
+    actualStartTime: 2887.6999999046326,
+    alternate: null,
+    child: subNode2,
+    childLanes: 0,
+    deletions: null,
+    dependencies: null,
+    elementType: {name: 'App'},
+    flags: 0,
+    index: 0,
+    key: null,
+    lanes: 0,
+    memoizedProps: null,
+    memoizedState: null,
+    mode: 3,
+    pendingProps: null,
+    ref: null,
+    refCleanup: null,
+    return: null,
+    selfBaseDuration: 0.09999990463256836,
+    sibling: null,
+    stateNode: null,
+    subtreeFlags: 14682117,
+    tag: 0,
+    treeBaseDuration: 5.699999809265137,
+    type: null,
+    updateQueue: null,
+    _debugHookTypes: null,
+    _debugNeedsRemount: false,
+    _debugOwner: null,
+    _debugSource: null,
+}
+
+
+const fiberNode = {
+    actualDuration: 0.5999999046325684,
+    actualStartTime: 2887.5999999046326,
+    alternate: null,
+    child: subNode,
+    childLanes: 0,
+    deletions: null,
+    dependencies: null,
+    elementType: null,
+    flags: 0,
+    index: 0,
+    key: null,
+    lanes: 0,
+    memoizedProps: null,
+    memoizedState: null,
+    mode: 3,
+    pendingProps: null,
+    ref: null,
+    refCleanup: null,
+    return: null,
+    selfBaseDuration: 0.09999990463256836,
+    sibling: null,
+    stateNode: null,
+    subtreeFlags: 14682117,
+    tag: 3,
+    treeBaseDuration: 5.699999809265137,
+    type: null,
+    updateQueue: null,
+    _debugHookTypes: null,
+    _debugNeedsRemount: false,
+    _debugOwner: null,
+    _debugSource: null,
+};
+
+module.exports = fiberNode;
